requirements.txt
APScheduler==3.11.2
deep-translator==1.11.4
Flask==3.1.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
gevent==24.10.3
gunicorn==21.2.0
numpy==2.4.3
openai-whisper==20250625
python-dotenv==1.0.1
SQLAlchemy==2.0.48
torch==2.10.0
Werkzeug==3.1.6
